requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 异步任务
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI 模型
openai-whisper==20231117
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
numpy<2.0.0

# 文本处理
textdistance==4.6.1
jieba==0.42.1
pydub==0.25.1

# 存储和网络
cos-python-sdk-v5==1.9.24
httpx==0.25.2

# 千问大模型API
dashscope==1.14.1

# 系统监控
psutil==5.9.6
prometheus-client==0.19.0

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1 